Generalize StorageBasedMatrix over backend
