+++ include/alp/amf-based/matrix.hpp
@@ -59,7 +59,7 @@ namespace alp {
 
 	namespace internal {
 		/** Forward declaration */
-		template< typename T, typename AmfType, bool requires_allocation >
+		template< typename T, typename AmfType, bool requires_allocation, Backend backend >
 		class StorageBasedMatrix;
 
 		/** Forward declaration */
@@ -182,6 +182,247 @@ namespace alp {
 				}
 
 		};
+
+		/**
+		 * Matrix container specialization
+		 * Implements both original containers and views on containers.
+		 * @tparam requires_allocation True if the class is an original container
+		 *                             False if the class is a view of another matrix
+		 */
+		template< typename T, typename AmfType, bool requires_allocation, Backend backend >
+		class StorageBasedMatrix : public MatrixBase< StorageBasedMatrix< T, AmfType, requires_allocation, backend > > {
+
+			template<
+				typename MatrixType,
+				std::enable_if_t< internal::is_storage_based< MatrixType >::value > *
+			>
+			friend size_t getStorageDimensions( const MatrixType &A ) noexcept;
+
+			/** Get the reference to the AMF of a storage-based matrix */
+			template<
+				typename MatrixType,
+				std::enable_if< internal::is_storage_based< MatrixType >::value > *
+			>
+			friend const typename MatrixType::amf_type &getAmf( const MatrixType &A ) noexcept;
+
+			public:
+
+				/** Expose static properties */
+
+				typedef T value_type;
+				typedef AmfType amf_type;
+				typedef typename AmfType::imf_r_type imf_r_type;
+				typedef typename AmfType::imf_c_type imf_c_type;
+				/** Type returned by access function */
+				typedef T &access_type;
+				/** Type of the index used to access the physical storage */
+				typedef size_t storage_index_type;
+
+			protected:
+				typedef StorageBasedMatrix< T, AmfType, requires_allocation, backend > self_type;
+				friend MatrixBase< self_type >;
+
+				typedef typename std::conditional<
+					requires_allocation,
+					Vector< T, backend >,
+					Vector< T, backend > &
+				>::type container_type;
+
+				/** A container-type view is characterized by its association with a physical container */
+				container_type container;
+
+				/**
+				 * All matrix views use a pair of index mapping functions to
+				 * capture the correspondence between their logical layout and the one
+				 * of their underlying container. This may be another view leading to a composition
+				 * of IMFs between the top matrix view and the physical container.
+				 * Original matrix containers's index mapping functions are an identity mapping.
+				 */
+				//ImfR imf_r;
+				//ImfC imf_c;
+
+				/**
+				 * The container's storage scheme. \a storage_scheme is not exposed to the user as an option
+				 * but can defined by ALP at different points in the execution depending on the \a backend choice.
+				 * In particular, if the structured matrix is not a temporary matrix than it is fixed at construction
+				 * time when the allocation takes place.
+				 * If the structured matrix is a temporary one than a storage storage scheme choice may or may not be
+				 * made depending on whether a decision about instantiating the matrix is made by the framework.
+				 *
+				 * The specific storage scheme choice depends on the chosen backend and the structure of the matrix.
+				 * \internal \todo Revisit this when introducing storage mapping functions.
+				 */
+				//Smf smf;
+
+				/**
+				 * Access mapping function maps a pair of logical coordinates
+				 * into the concrete coordinate inside the actual container.
+				 * \see AMF
+				 */
+				AmfType amf;
+				/**
+				 * @brief determines the size of the matrix via the domain of
+				 * the index mapping functions.
+				 *
+				 * @return A pair of dimensions.
+				 */
+				std::pair< size_t, size_t > dims() const noexcept {
+					return amf.getLogicalDimensions();
+				}
+
+				size_t getStorageDimensions() const noexcept {
+					return amf.getStorageDimensions();
+				}
+
+				friend const Vector< T, backend > & getContainer( const self_type &A ) {
+					return A.container;
+				}
+
+				friend Vector< T, backend > & getContainer( self_type &A ) {
+					return A.container;
+				}
+
+				bool getInitialized() const noexcept {
+					return internal::getInitialized( container );
+				}
+
+				void setInitialized( const bool initialized ) noexcept {
+					internal::setInitialized( container , initialized );
+				}
+
+				const AmfType &getAmf() const noexcept {
+					return amf;
+				}
+
+				/**
+				 * Returns a constant reference to the element corresponding to
+				 * the provided storage index.
+				 *
+				 * @param storageIndex  storage index in the physical iteration
+				 *                      space.
+				 *
+				 * @return const reference or value of the element at given position.
+				 */
+				const access_type access( const storage_index_type &storageIndex ) const {
+					return container[ storageIndex ];
+				}
+
+				access_type access( const storage_index_type &storageIndex ) {
+					return container[ storageIndex ];
+				}
+
+				storage_index_type getStorageIndex( const size_t i, const size_t j, const size_t s, const size_t P ) const {
+					return amf.getStorageIndex( i, j, s, P );
+				}
+
+				/**
+				 * @brief Construct a new structured matrix Base object assigning identity
+				 * mapping functions both to the row and column dimensions.
+				 *
+				 * @param rows The number of rows of the matrix.
+				 * @param cols The number of columns of the matrix.
+				 * @param smf  The storage mapping function assigned to this matrix.
+				 */
+				/** (Documentation of old MatrixCotiainer. TODO: merge with the above docs.
+				 * @brief Construct a new structured matrix container object.
+				 *
+				 * \warning \a cap is present for compatibility with other matrix specializations.
+				 *          In reference backend, the number of non-zeros (i.e. capacity)
+				 *          depends on the used storage scheme. Therefore, this parameter is
+				 *          ignored.
+				 *
+				 * TODO: Add the storage scheme a parameter to the constructor
+				 * so that allocation can be made accordingly, generalizing the full case.
+				 */
+				StorageBasedMatrix( AmfType &&amf ) :
+					// enable only if ImfR and ImfC are imf::Id
+					container( internal::Vector< T, backend >( amf.getStorageDimensions() ) ),
+					amf( std::move( amf ) ) {}
+
+				/** View on another container */
+				StorageBasedMatrix( Vector< T, backend > &container, AmfType &&amf ) :
+					container( container ),
+					amf( std::move( amf ) ) {}
+
+		}; // class StorageBasedMatrix
+
+		/**
+		 * Specialization of MatrixReference with a lambda function as a target.
+		 * Used as a result of low-rank operation to avoid the need for allocating a container.
+		 * The data is produced lazily by invoking the lambda function stored as a part of this object.
+		 *
+		 * \note Views-over-lambda-functions types are used internally as results of low-rank operations and are not
+		 *       directly exposed to users. From the users perspective, the use of objects of this type does not differ
+		 *       from the use of other \a alp::Matrix types. The difference lies in a lazy implementation of the access
+		 *       to matrix elements, which is not exposed to the user.
+		 *
+		 */
+		template< typename T, typename ImfR, typename ImfC, typename DataLambdaType >
+		class FunctorBasedMatrix : public MatrixBase< FunctorBasedMatrix< T, ImfR, ImfC, DataLambdaType > > {
+			public:
+
+				/** Expose static properties */
+				typedef T value_type;
+				/** Type returned by access function */
+				typedef T access_type;
+				/** Type of the index used to access the physical storage */
+				typedef std::pair< size_t, size_t > storage_index_type;
+
+			protected:
+
+				typedef FunctorBasedMatrix< T, ImfR, ImfC, DataLambdaType > self_type;
+				friend MatrixBase< self_type >;
+
+				typedef std::function< bool() > initialized_functor_type;
+				const initialized_functor_type initialized_lambda;
+
+				const ImfR imf_r;
+				const ImfC imf_c;
+
+				const DataLambdaType data_lambda;
+
+				std::pair< size_t, size_t > dims() const noexcept {
+					return std::make_pair( imf_r.n, imf_c.n );
+				}
+
+				const DataLambdaType &getFunctor() const noexcept {
+					return data_lambda;
+				}
+
+				bool getInitialized() const noexcept {
+					return initialized_lambda();
+				}
+
+				void setInitialized( const bool ) noexcept {
+					static_assert( "Calling setInitialized on a FunctorBasedMatrix is not allowed." );
+				}
+
+				access_type access( const storage_index_type &storage_index ) const {
+					T result = 0;
+					data_lambda( result, imf_r.map( storage_index.first ), imf_c.map( storage_index.second ) );
+					return static_cast< access_type >( result );
+				}
+
+				storage_index_type getStorageIndex( const size_t i, const size_t j, const size_t s, const size_t P ) const {
+					(void)s;
+					(void)P;
+					return std::make_pair( i, j );
+				}
+
+			public:
+
+				FunctorBasedMatrix(
+					initialized_functor_type initialized_lambda,
+					ImfR imf_r,
+					ImfC imf_c,
+					const DataLambdaType data_lambda
+				) :
+					initialized_lambda( initialized_lambda ),
+					imf_r( imf_r ),
+					imf_c( imf_c ),
+					data_lambda( data_lambda ) {}
+
+		}; // class FunctorBasedMatrix
 
 		/**
 		 * Determines the mapping polynomial type and exposes a factory method
@@ -314,7 +555,8 @@ namespace alp {
 				internal::FunctorBasedMatrix< T, ImfR, ImfC, typename View::applied_to >,
 				internal::StorageBasedMatrix< T,
 					typename internal::determine_amf_type< Structure, View, ImfR, ImfC, backend >::type,
-					internal::requires_allocation< View >::value
+					internal::requires_allocation< View >::value,
+					backend
 				>
 			>::type type;
 		};
